refactor: :recycle: move various shared types to global types file
